pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sanna"
version = "0.12.0"
description = "Coherence checks for AI agents — portable, offline-verifiable reasoning receipts"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [{name = "nicallen-exd"}]
keywords = ["ai", "governance", "reasoning", "receipts", "verification", "observability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "jsonschema>=4.17",
    "pyyaml>=6.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
langfuse = ["langfuse>=2.0"]
mcp = ["mcp>=1.0"]
otel = ["opentelemetry-api>=1.20.0", "opentelemetry-sdk>=1.20.0"]
dev = ["pytest>=7.0", "jsonschema>=4.17", "pyyaml>=6.0"]

[project.scripts]
sanna-generate = "sanna.cli:main_generate"
sanna-verify = "sanna.cli:main_verify"
sanna-init-constitution = "sanna.cli:main_init_constitution"
sanna-sign-constitution = "sanna.cli:main_sign_constitution"
sanna-keygen = "sanna.cli:main_keygen"
sanna-hash-constitution = "sanna.cli:main_hash_constitution"
sanna-verify-constitution = "sanna.cli:main_verify_constitution"
sanna-mcp = "sanna.mcp.__main__:main"
sanna-create-bundle = "sanna.cli:main_create_bundle"
sanna-verify-bundle = "sanna.cli:main_verify_bundle"
sanna-drift-report = "sanna.cli:main_drift_report"
sanna-approve-constitution = "sanna.cli:approve_constitution_cmd"
sanna-diff = "sanna.cli:diff_cmd"
sanna-init = "sanna.init_constitution:main"
sanna-gateway = "sanna.gateway:main"
# Legacy aliases
c3m-receipt = "sanna.cli:main_generate"
c3m-verify = "sanna.cli:main_verify"

[project.urls]
Homepage = "https://github.com/nicallen-exd/sanna"
Issues = "https://github.com/nicallen-exd/sanna/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sanna = ["spec/*.json", "templates/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
